Virheidenhallinta jonkinmoiseen kuntoon. Myös clientin puolelta. Testien korjaaminen

+++ server/controllers/products.js
@@ -3,7 +3,7 @@ const Product = require('../models/product')
 const Instruction = require('../models/instruction')
 const authUtils = require('../utils/auth')
 const STATUS_CODES = require('http-status')
-const { ParameterValidationException, buildValidationErrorObjectFromMongooseValidationError, ResourceNotFoundException } = require('../error/exceptions')
+const { ResourceNotFoundException, restructureCastAndValidationErrorsFromMongoose, UnauthorizedException } = require('../error/exceptions')
 
 productRouter.get('/', async (req, res, next) => {
   try {
@@ -14,9 +14,11 @@ productRouter.get('/', async (req, res, next) => {
     })
 
     products.forEach(p => p.instructions.sort((a, b) => b.score - a.score))
-    res.json(products.map((product) => product.toJSON()))
+    res.status(STATUS_CODES.OK).json(products.map((product) => product.toJSON()))
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
@@ -28,9 +30,11 @@ productRouter.get('/user', async (req, res, next) => {
       user: 1
     })
     favorites.forEach(p => p.instructions.sort((a, b) => b.score - a.score))
-    res.json(favorites.map((favorite) => favorite.toJSON()))
+    res.status(STATUS_CODES.OK).json(favorites.map((favorite) => favorite.toJSON()))
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
@@ -46,10 +50,12 @@ productRouter.get('/:id', async (req, res, next) => {
     }
 
     product.instructions.sort((a, b) => b.score - a.score)
-    res.json(product)
+    res.status(STATUS_CODES.OK).json(product)
 
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
@@ -58,21 +64,16 @@ productRouter.post('/', async (req, res, next) => {
     let user = await authUtils.authenticateRequestReturnUser(req)
 
     const body = req.body
-    try {
-      const product = new Product({
-        name: body.name,
-        user: user.id,
-      })
-      const result = await product.save()
-      return res.status(STATUS_CODES.CREATED).json(result)
-    } catch (error) {
-      // return res.status(STATUS_CODES.BAD_REQUEST).
-      //   json({ error: 'product name required' })
-      let validationErrorObject = buildValidationErrorObjectFromMongooseValidationError(error)
-      throw new ParameterValidationException('Tuotteen nimi vaaditaan!', validationErrorObject)
-    }
+    const product = new Product({
+      name: body.name,
+      user: user.id,
+    })
+    const result = await product.save()
+    return res.status(STATUS_CODES.CREATED).json({ message: 'Tuote luotu onnistuneesti!', resource: result })
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
@@ -82,7 +83,7 @@ productRouter.post('/:id/instructions', async (req, res, next) => {
 
     const product = await Product.findById(req.params.id)
     if (!product) {
-      return res.status(STATUS_CODES.UNAUTHORIZED).json({ error: 'No product' })
+      throw ResourceNotFoundException('Tuotetta ID:llä: ' + req.params.id + ' ei löytynyt!')
     }
     const instruction = new Instruction(req.body)
     instruction.product = product.id
@@ -90,41 +91,45 @@ productRouter.post('/:id/instructions', async (req, res, next) => {
     const result = await instruction.save()
     product.instructions = product.instructions.concat(result)
     await product.save()
-    res.status(STATUS_CODES.CREATED).json(result)
+    res.status(STATUS_CODES.CREATED).json({ message: 'Ohje luotu onnistuneesti!', resource: result})
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
+/**
+ * Poistaa tuotteeseen liittyvän ohjeen
+ */
 productRouter.delete('/:productId/instructions/:instructionId', async (req, res, next) => {
   try {
     let user = await authUtils.authenticateRequestReturnUser(req)
 
-    let product    //haetaan tuote, johon ohje liittyy
-    try {
-      product = await Product.findById(req.params.productId)
-    } catch (e) {
-      return res.status(STATUS_CODES.UNAUTHORIZED).json({ error: e.message })
-    }
+    let product = await Product.findById(req.params.productId)
 
-    let instruction   //haetaan ohje
-    try {
-      instruction = await Instruction.findById(req.params.instructionId)
-    } catch (e) {
-      return res.status(STATUS_CODES.BAD_REQUEST).json({ error: e.message })
-    }
+    let instruction = await Instruction.findById(req.params.instructionId)
 
     //verrataan, vastaako pyynnön tehnyt käyttäjä ohjeen lisännyttä käyttäjää
     if (instruction.user.toString() !== user.id.toString()) {
-      return res.status(STATUS_CODES.FORBIDDEN).json({ error: 'unauthorized'})
+      throw new UnauthorizedException('Vain ohjeen luoja voi poistaa ohjeen!')
     }
 
+    let deletedInstructionText = instruction.information
     //poistetaan ohje tietokannasta
     product.instructions = product.instructions.pull({ _id: instruction.id })
     await product.save()
-    res.status(STATUS_CODES.CREATED).json(product)
+    res.status(STATUS_CODES.OK)
+      .json(
+        {
+          message: 'Ohje \'' + deletedInstructionText + '\' poistettiin onnistuneesti!',
+          resource: product
+        }
+      )
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
 
@@ -137,27 +142,23 @@ productRouter.delete('/:id', async (req, res, next) => {
   try {
     let user = await authUtils.authenticateRequestReturnUser(req)
 
-    let product   // haetaan poistettava tuote
-    try {
-      product = await Product.findById(req.params.id).exec()
-    } catch (error) {
-      return res.status(STATUS_CODES.UNAUTHORIZED).json({ error: error.message })
+    let product = await Product.findById(req.params.id).exec()
+    if (!product) {
+      throw new ResourceNotFoundException('Tuotetta ID:llä: ' + req.query.productID + ' ei löytynyt!')
     }
 
     // verrataan pyynnön tehnyttä käyttäjää tuotteen lisänneeseen käyttäjään
     if (product.user.toString() !== user.id.toString()) {
-      return res.status(STATUS_CODES.FORBIDDEN).json({ error: 'unauthorized' })
+      throw new UnauthorizedException('Vain tuotteen luoja voi poistaa ohjeen!')
     }
-
+    let deletedProductName = product.name
     // poistetaan tuote tietokannasta
-    try {
-      await Product.findByIdAndDelete({ _id: req.params.id }).exec()
-      res.status(STATUS_CODES.OK).json()
-    } catch (error) {
-      return res.status(STATUS_CODES.BAD_REQUEST).json({ error: error.message })
-    }
+    await Product.findByIdAndDelete({ _id: req.params.id }).exec()
+    res.status(STATUS_CODES.OK).json({message: 'Tuote \'' + deletedProductName + '\' poistettiin onnistuneesti!'})
   } catch (error) {
-    next(error)
+    let handledError = restructureCastAndValidationErrorsFromMongoose(error)
+    // To the errorhandler in app.js
+    next(handledError)
   }
 })
